Modified lighthouse-badges file
