+++ lighthouserc.json
@@ -1,13 +1,21 @@
 {
+  "extends": "lighthouse:default",
+  "siteTitle": "Lighthouse Report",
+
   "ci": {
     "collect": {
-      "numberOfRuns": 3
+      "numberOfRuns": 1,
+      "url": ["http://localhost:8080/"],
+      "startServerCommand": "npm run start:static"
     },
     "assert": {
-      "preset": "lighthouse:recommended",
       "assertions": {
-        "offscreen-images": "off",
-        "uses-webp-images": "off"
+        "first-contentful-paint": "off",
+        "categories:performance": ["error", {"minScore":  0}],
+        "categories:accessibility": ["error", {"minScore":  0.9}],
+        "categories:best-practices": ["error", {"minScore":  0}],
+        "categories:seo": ["error", {"minScore":  0}],
+        "categories:pwa":["error", {"minScore":  0}]
       }
     },
     "upload": {
