+++ package.json
@@ -25,7 +25,7 @@
     "start": "ng serve --port 4200",
     "start:static": "node scripts/serve-static.js",
     "test": "npm run test:unit && npm run test:e2e",
-    "test:audit": "node scripts/build-wvr-components-lighthouse.js",
+    "test:audit": "rimraf .lighthouseci && lhci autorun --upload.target=temporary-public-storage --config=./lighthouserc.json && node scripts/build-wvr-components-lighthouse-badges.js",
     "test:e2e": "ng e2e",
     "test:unit": "ng test --no-watch",
     "test:watch": "ng test",
@@ -34,11 +34,10 @@
     "build:docs": "npm run build:docs-usage && npm run build:docs-development",
     "build:docs-development": "compodoc -d static/docs/development -p ./projects/wvr-elements/tsconfig.docs.json",
     "build:docs-usage": "node scripts/build-wvr-ud.js",
-    "build:coverage-report": "cat ./coverage/wvr-elements/lcov.info | ./node_modules/coveralls/bin/coveralls.js",
-    "build:static": "npm run build:static-setup && npm run build:static-docs && npm run build:static-reports && npm run build:coverage-report",
+    "build:static": "npm run build:static-docs && npm run build:static-reports",
     "build:static-setup": "node scripts/build-wvr-components-static.js",
     "build:static-docs": "npm run build:static-setup && npm run build:docs-usage && compodoc -d static/docs/development -p ./projects/wvr-elements/tsconfig.docs.json",
-    "build:static-reports": "npm run build:static-setup && npm run test:audit && npm run test:coverage",
+    "build:static-reports": "npm run build:static-setup && cat ./static/reports/coverage/wvr-elements/lcov.info | ./node_modules/coveralls/bin/coveralls.js",
     "build:static-production": "npm run build:static && ls -A | grep -v static | xargs rimraf && mv static/* . && rimraf static",
     "static:serve": "static-server static -p 8081",
     "static:clean": "rimraf static"
@@ -54,6 +53,7 @@
     "@angular/platform-browser": "^9.0.1",
     "@angular/platform-browser-dynamic": "^9.0.1",
     "@angular/router": "^9.0.1",
+    "@lhci/cli": "^0.3.14",
     "@ng-bootstrap/ng-bootstrap": "^5.3.0",
     "@storybook/react": "^5.3.18",
     "bootstrap": "^4.4.1",
@@ -67,6 +67,7 @@
     "glob": "^7.1.6",
     "ie11-custom-properties": "^3.0.6",
     "jsdom": "^16.2.2",
+    "lighthouse-badges": "^1.0.27",
     "rxjs": "~6.5.4",
     "tslib": "^1.10.0",
     "web-animations-js": "^2.3.2",
@@ -106,6 +107,6 @@
     "tsickle": "^0.38.0",
     "tslint": "~6.0.0",
     "tslint-angular": "^3.0.2",
-    "typescript": "3.6.4"
+    "typescript": ">=3.6.4 and <= 3.9.0"
   }
 }
