+++ projects/wvr-elements/karma.conf.js
@@ -16,7 +16,7 @@ module.exports = function (config) {
       clearContext: false // leave Jasmine Spec Runner output visible in browser
     },
     coverageIstanbulReporter: {
-      dir: require('path').join(__dirname, '../../static/report/coverage/wvr-elements'),
+      dir: require('path').join(__dirname, '../../static/reports/coverage/wvr-elements'),
       reports: ['html', 'lcovonly', 'text-summary'],
       fixWebpackSourcePaths: true
     },
